requirements.txt
pydantic_settings
pydantic
ultralytics
ray
tqdm
fastapi
uvicorn
numpy
optuna
optuna-dashboard
Cython
h5py
Pillow
six
scipy
opencv-python
matplotlib
tb-nightly
future
yacs
gdown
flake8
yapf
isort==4.3.21
imageio
chardet
# Export --------------------------------------
# onnx
# onnx-simplified
# openvino-dev
# openvino2tensorflow
# tensorflow
# tensorflow_datasets
